Created the Todo.findAll method
