+++ .mocharc.json
@@ -1,0 +1,3 @@
+{
+  "recursive": true
+}
