--- Todo.js
@@ -1,16 +1,0 @@
-const todos = [];
-let nextId = 1;
-
-class Todo {
-  constructor({title, body}) {
-    this.id = nextId++;
-    this.title = title;
-    this.body = body;
-    this.createdAt = new Date();
-    this.updatedAt = new Date();
-  }
-}
-
-module.exports = {
-
-};
