+++ package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "mocha"
   },
   "repository": {
     "type": "git",
@@ -15,5 +15,9 @@
   "bugs": {
     "url": "https://github.com/gorinesia/practice-express.js/issues"
   },
-  "homepage": "https://github.com/gorinesia/practice-express.js#readme"
+  "homepage": "https://github.com/gorinesia/practice-express.js#readme",
+  "devDependencies": {
+    "mocha": "^8.1.3",
+    "power-assert": "^1.6.1"
+  }
 }
